update/ split GroupItem
