+++ modules/UserCenter/MyGroups/GroupsList.tsx
@@ -1,20 +1,8 @@
 import React, { useState } from "react";
-import Link from "@/common/components/GeneralLink";
-import Button from "@/common/components/GeneralButton";
-import TabSection from "../TabSection";
-import formatDate from "@/common/helpers/formateDate";
-import statusSet from "@/constants/groupStatusSet";
-import { groupsData, tabs, ActionBtnType } from "./data";
 import { useRouter } from "next/router";
-
-// 是哪一種狀態
-const setStatus = (endTime: string, status: string) => {
-  const now = new Date();
-  const today = now.toISOString();
-  if (status === "pending") return "pending";
-  if (endTime < today) return "over";
-  return "member";
-};
+import TabSection from "../TabSection";
+import GroupItem from "./GroupItem";
+import { groupsData, tabs, ActionBtnsType } from "./data";
 
 export default function GroupsList({}) {
   const router = useRouter();
@@ -26,7 +14,7 @@ export default function GroupsList({}) {
     router.push("/about");
   };
 
-  const actionBtns: Record<string, ActionBtnType> = {
+  const actionBtns: ActionBtnsType = {
     pending: {
       text: "取消申請",
       func: handleQuitGroup,
@@ -79,90 +67,13 @@ export default function GroupsList({}) {
         </div>
         <ul>
           {groupsData.map((group) => {
-            const {
-              groupId,
-              groupName,
-              store,
-              place,
-              totalMemberNum,
-              currentNum,
-              startTime,
-              endTime,
-              status,
-              commented,
-            } = group;
-
-            // 得到狀態："pending" "member" "over"
-            const groupStatus = setStatus(endTime, status);
-            // 狀態表示
-            const statusStyle = statusSet[groupStatus].style;
-            const statusText = statusSet[groupStatus].text;
-
-            const isActive =
-              (isActiveOver && groupStatus === "over") ||
-              (!isActiveOver && groupStatus !== "over");
-            if (!isActive) return;
-
-            // ----
-
-            // 目前的評價狀態如何，會從 API 得取
-            const isCommented = commented;
-
-            // 自動抓取對應的按鈕
-            //如果是 activeTab === over，看是否已有評價？
-            const selectBtn = () => {
-              if (!isActiveOver) return groupStatus;
-              if (isCommented) return "commented";
-              return "over";
-            };
-            const actionBtnId = selectBtn();
-            const btnDisabled = actionBtns[actionBtnId].disabled;
-            const btnOnClick = actionBtns[actionBtnId].func;
-            const btnText = actionBtns[actionBtnId].text;
-
-            const isPlace = place === "NULL";
-            const location = isPlace ? store.storeName : place;
-
-            const { groupDate, formatStartTime, formatEndTime } = formatDate(
-              startTime,
-              endTime
-            );
-            const groupTime = `${formatStartTime} - ${formatEndTime}`;
-
             return (
-              <li
-                key={groupId}
-                className="w-full flex justify-between items-center p-2 mb-3 bg-yellow-tint text-center text-sm"
-              >
-                <p className="w-[10%] text-xs">
-                  <span className={`groupStatusDot ${statusStyle}`}>
-                    {statusText}
-                  </span>
-                </p>
-                <p className="w-[20%] truncate text-sm">
-                  <Link href={`/group/${groupId}`}>{groupName}</Link>
-                </p>
-                <p className="w-[20%] truncate">{location}</p>
-                <p className="w-[20%] flex flex-wrap justify-center gap-2">
-                  <span className=" whitespace-nowrap">{groupDate}</span>
-                  <span className=" whitespace-nowrap">{groupTime}</span>
-                </p>
-                <p className="w-[10%]">
-                  {currentNum}/{totalMemberNum}
-                </p>
-                <div className="w-[10%]">
-                  <Button
-                    type="button"
-                    appearance="black"
-                    rounded
-                    className="w-full"
-                    isDisabled={btnDisabled}
-                    onClick={btnOnClick}
-                  >
-                    <p className="text-sm">{btnText}</p>
-                  </Button>
-                </div>
-              </li>
+              <GroupItem
+                key={group.groupId}
+                group={group}
+                isActiveOver={isActiveOver}
+                actionBtns={actionBtns}
+              />
             );
           })}
         </ul>
